+++ Source/LoginServer/LoginServer/DataManager.cpp
@@ -23,7 +23,7 @@ NativeInfo::Stat DataManager::GetCharacterCreateStat(const Define::Job& job)
 	return this->characterCreateStat[job];
 }
 
-void DataManager::GetCharacterCreateGear(const Define::Job& job, std::array<uint8_t, Define::GearType_END>& out)
+void DataManager::GetCharacterCreateGear(const Define::Job& job, std::array<int32_t, Define::GearType_END>& out)
 {
 	std::copy(this->characterCreateGear[job].index.begin(), this->characterCreateGear[job].index.end(), out.begin());
 }
